+++ cirq-google/cirq_google/workflow/quantum_runtime_test.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 import glob
 import re
+import time
 import uuid
 from typing import List, cast, Any
 
@@ -22,6 +23,7 @@ import pytest
 import cirq
 import cirq_google as cg
 from cirq_google.workflow.quantum_executable_test import _get_quantum_executables, _get_example_spec
+from cirq_google.workflow.quantum_runtime import _time_into_runtime_info
 
 
 def cg_assert_equivalent_repr(value):
@@ -107,6 +109,15 @@ def test_executable_group_result(tmpdir):
     cg_assert_equivalent_repr(egr)
     assert len(egr.executable_results) == 3
     _assert_json_roundtrip(egr, tmpdir)
+
+
+def test_timing():
+    rt = cg.RuntimeInfo(execution_index=0)
+    with _time_into_runtime_info(rt, 'test_proc'):
+        time.sleep(0.1)
+
+    assert 'test_proc' in rt.timings_s
+    assert rt.timings_s['test_proc'] > 0.05
 
 
 def _load_result_by_hand(tmpdir: str, run_id: str) -> cg.ExecutableGroupResult:
