--- cirq-google/cirq_google/workflow/timing_test.py
@@ -1,25 +1,0 @@
-# Copyright 2022 The Cirq Developers
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     https://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-import time
-
-import cirq_google as cg
-
-
-def test_timing():
-    rt = cg.RuntimeInfo(execution_index=0)
-    with cg.workflow.TimeIntoRuntimeInfo(rt, 'test_proc'):
-        time.sleep(0.1)
-
-    assert 'test_proc' in rt.timings_s
-    assert rt.timings_s['test_proc'] > 0.05
